@@ -15,7 +15,6 @@ classifiers = [
     "Intended Audience :: Developers",
     "License :: OSI Approved :: MIT License",
     "Programming Language :: Python :: 3",
-    "Programming Language :: Python :: 3.8",
     "Programming Language :: Python :: 3.9",
     "Programming Language :: Python :: 3.10",
     "Programming Language :: Python :: 3.11",
@@ -34,7 +33,7 @@ packages = [{include = "db_types", from = "src"}]
 "Changelog" = "https://github.com/gurmeetsaran/python-db-types/blob/main/CHANGELOG.md"
 
 [tool.poetry.dependencies]
-python = "^3.8"
+python = "^3.9"
 typing-extensions = {version = ">=4.0.0", python = "<3.9"}
 
 [tool.poetry.group.dev.dependencies]
@@ -56,7 +55,7 @@ pydantic = "^2.0.0"
 
 [tool.ruff]
 line-length = 100
-target-version = "py38"
+target-version = "py39"
 
 [tool.ruff.lint]
 select = [
@@ -75,8 +74,11 @@ select = [
 [tool.ruff.lint.isort]
 known-first-party = ["db_types"]
 
+[tool.ruff.lint.per-file-ignores]
+"src/db_types/annotations.py" = ["N802"]  # Allow PascalCase function names for clean API
+
 [tool.pyright]
-pythonVersion = "3.8"
+pythonVersion = "3.9"
 typeCheckingMode = "basic"
 include = ["src"]
 exclude = ["**/__pycache__", "**/.pytest_cache", "venv", ".venv", "tests", "examples"]
@@ -111,7 +113,7 @@ fail_under = 59
 
 [tool.black]
 line-length = 100
-target-version = ["py38"]
+target-version = ["py39"]
 include = '\.pyi?$'
 
 [tool.isort]
